refactored structure for clarity
