+++ imu/imu_test.py
@@ -1,0 +1,4 @@
+from BerryIMU import process_imu_data
+
+if __name__ == '__main__':
+    process_imu_data()
